code-revision: drastic structural changes of the codebase
- The only what is missing, is `README.md` to address the current exposed implementation, hence "documentation"
