+++ src/views/htmlcanvas/DOMutils.js
