+++ src/views/htmlcanvas/modules/maths/index.js
