+++ src/views/xmlsvg/DOMutils.js
@@ -1,0 +1,4 @@
+globalThis.ArgsList = Array;
+
+EventTarget.prototype.on = EventTarget.prototype.addEventListener;
+EventTarget.prototype.rm = EventTarget.prototype.removeEventListener;
