+++ src/views/xmlsvg/svg-circle/index.js
@@ -1,4 +1,4 @@
-import { getNamespace } from "../utils.js";
+import { getNamespace } from "../modules/index.js";
 
 export const svg_circle = getNamespace(import.meta.url);
 customElements.define(svg_circle, class extends HTMLElement {
