+++ src/views/xmlsvg/svg-path/index.js
@@ -1,4 +1,4 @@
-import { getNamespace } from "../utils.js";
+import { getNamespace } from "../modules/index.js";
 
 export const svg_path = getNamespace(import.meta.url);
 customElements.define(svg_path, class extends HTMLElement {
