+++ src/views/xmlsvg/svg-rect/index.js
@@ -1,4 +1,4 @@
-import { getNamespace } from "../utils.js";
+import { getNamespace } from "../modules/index.js";
 
 export const svg_rect = getNamespace(import.meta.url);
 customElements.define(svg_rect, class extends HTMLElement {
